Add Travis-CI style and build check
Add a skeleton .travis.yml file that runs the code style
checker and builds the project. This can be later extended to
run a test suite. A UnifyCR project maintainer just needs to
log in to http://travis-ci.org and enable testing for the
LLNL/UnifyCR repo. (Once the repo is made public, that is.
While LLNL/UnifyCR is still private we would have to use
http://travis-ci.com instead.)

Update checkpatch.sh to use 'git format-patch' instead of
'git show' to produce pactches for checkpatch.pl. The output
of format-patch is more in line with what checkpatch.pl expects.
In particular, commit messages are not indented and it allows
one to check a whole patch stack by specifying a revision range,
e.g.: ./scripts/checkpatch.sh origin/dev..HEAD

Remove redundant comment in checkpatch.sh.