Make rm_delegate_request() a valid pthread start routine
Update rm_delegate_request() so its type matches that required by
pthread_create():

  void *(*start_routine) (void *)

The former type mismatch triggered a compiler warning:

  In file included from unifycr_global.h:32:0,
                   from unifycr_cmd_handler.c:37:
  /usr/include/pthread.h:244:12: note: expected void * (*)(void *)
  but argument is of type int (*)(struct cli_signature_t *)
   extern int pthread_create (pthread_t *__restrict __newthread,
            ^

Rename the function to rm_delegate_request_thread() to make it more
clear that it's invoked as a thread start routine. The former
success/failure int return values were changed to NULL. Nothing
in-tree appears to be making use of the return values, so this should
be safe. If rm_delegate_request() was meant to also serve as a
library function that returns an integer error code, it should
probably be implemented separately from the pthread start routine.