Add rudimentary functional test suite
Add a test suite based on sharness, libtap, and automake. Use
sharness and libtap to implement tests in shell scripts and C
programs, respectively, that produce Test Anything Protocol (TAP)
output.  Integrate with the automake test driver so the suite can be
run with 'make check' or 'make distcheck'.

The current tests are not designed to be run in parallel, i.e.  with
'make -jN'. In particular, the tests must be run in the right order to
start and stop unifycrd at the appropriate time, but this dependency is
not expressed in the Makefile. This could be done by expressing Makefile
depdencies for the corresponding test log files, but this might create
some clutter in the Makefile rules if not done carefully. We'll want to
revisit this in the future.

Include an example test program called sys/open.t that exercises a
few open() and close() system calls, and test scripts to start and
stop unifycrd. The t/sys directory will eventually contain test
coverage for other wrapped system calls.

The current test suite is single-node and single-process for both the
UnifyCR server and client applications. A great deal of basic test
coverage can be achieved in this manner. Future development work
might allow multi-node/multi-process environments to be simulated on
a single node by using unique per-unifycrd metadata directories. For
test coverage in true multi-node environements, we will need to
integrate the test suite with a distrubited execution framework.

The results currently look like this:

  $ make check
  ...
  make  check-TESTS
  make[2]: Entering directory `/usr/workspace/wsb/bass6/unifycr/t'
  make[3]: Entering directory `/usr/workspace/wsb/bass6/unifycr/t'
  PASS: 0001-setup.t 1 - unifycrd running
  PASS: 0100-sys-open.t 1 - unifycr_mount at /var/tmp/bass6/tmp.9tsnWBqReP
    (rc=0)
  PASS: 0100-sys-open.t 2 - open non-existing file
    /var/tmp/bass6/tmp.9tsnWBqReP/Lxd1t1CJ@F+7mlSEuFFhE9D-wgpGdycpg flags
    O_CREAT|O_EXCL (fd=1025): Success
  PASS: 0100-sys-open.t 3 - close
    /var/tmp/bass6/tmp.9tsnWBqReP/Lxd1t1CJ@F+7mlSEuFFhE9D-wgpGdycpg (rc=0):
    Success
  PASS: 0100-sys-open.t 4 - open existing file
    /var/tmp/bass6/tmp.9tsnWBqReP/Lxd1t1CJ@F+7mlSEuFFhE9D-wgpGdycpg
    O_CREAT|O_EXCL should fail (fd=-1): File exists
  PASS: 0100-sys-open.t 5 - open existing file
    /var/tmp/bass6/tmp.9tsnWBqReP/Lxd1t1CJ@F+7mlSEuFFhE9D-wgpGdycpg O_RDWR
    (fd=1025): Success
  PASS: 0100-sys-open.t 6 - close
    /var/tmp/bass6/tmp.9tsnWBqReP/Lxd1t1CJ@F+7mlSEuFFhE9D-wgpGdycpg (rc=0):
    Success
  PASS: 9005-unifycr-unmount.t 1 - unifycr_mount at
    /var/tmp/bass6/tmp.9tsnWBqReP (rc=0)
  PASS: 9005-unifycr-unmount.t 2 - unifycr_unmount succeeds (rc=0)
  PASS: 9010-stop-unifycrd.t 1 - Stop unifycrd
  PASS: 9020-mountpoint-empty.t 1 - Intercepted mount point
    /var/tmp/bass6/tmp.9tsnWBqReP is empty
  make[4]: Entering directory `/usr/workspace/wsb/bass6/unifycr/t'
  Making all in lib
  make[5]: Entering directory `/usr/workspace/wsb/bass6/unifycr/t/lib'
  make[5]: Nothing to be done for `all'.
  make[5]: Leaving directory `/usr/workspace/wsb/bass6/unifycr/t/lib'
  make[5]: Entering directory `/usr/workspace/wsb/bass6/unifycr/t'
  make[5]: Nothing to be done for `all-am'.
  make[5]: Leaving directory `/usr/workspace/wsb/bass6/unifycr/t'
  make[4]: Leaving directory `/usr/workspace/wsb/bass6/unifycr/t'
  =========================================================================
  Testsuite summary for unifycr 0.1.0-94-gd8ef6d4
  =========================================================================
  # TOTAL: 11
  # PASS:  11
  # SKIP:  0
  # XFAIL: 0
  # FAIL:  0
  # XPASS: 0
  # ERROR: 0
  =========================================================================

- Marking this commit with

TEST_CHECKPATCH_SKIP_FILES=t/sharness.sh,t/lib/tap.h,t/lib/tap.c
TEST_CHECKPATCH_SKIP_FILES+=,t/tap-driver.sh

  due to style failures in sharness.sh, tap-driver.sh and libtap
  which are provided as-is by their respective external projects.

- Marking with

TEST_CHECKPATCH_ALLOW_FAILURE=yes

  to override these warnings which aren't applicable in this case.

  WARNING: break quoted strings at a space character
  #741: FILE: t/lib/testutil.c:40:
  +    const char charset[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
  +                           "abcdefghijklmnopqrstuvwxyz"

  WARNING: break quoted strings at a space character
  #742: FILE: t/lib/testutil.c:41:
  +                           "abcdefghijklmnopqrstuvwxyz"
  +                           "0123456789"