Use HOST_NAME_MAX as hostname buffer size in unifycr.h
Problem: unifycr.h uses a constant defined in unifycr-defs.h
but doesn't include that header because it's private.
This causes a compilation error for C files that include
the header unifycr.h which defines the  public API.

Solution: use HOST_NAME_MAX, which is defined in limits.h,
to set the size of the hostname buffer, rather than
UNIFYCR_MAX_FILENAME which is not publicly visible.

Fixes #23